fix: Convert String to Hashed Key on Compile-time
In the old and temporary way, users had got to get a hashed key for 'VariableTable' from the separate tool.  That was bothersome, and a counterintuitive hased key didn't tell what it means.
Now, thanks to keyword 'constexpr', users can be on high level with a meaningful string, and still this game program manipulates the secure hashed key.

etc: The sequence on start can be specified, in other words, one or more sequences can be skipped for the convenience in development.  Open the script '_OnlyDUringDev.lua'( the name can be revised. )
Textures' width and height are set to the power of 2, the video-ram-friendly number.

( Through the blank term between the last commit and this, I could assure that the approach to the logics, especially sequence transition and global variable management is not strange or weird by using Unity engine. )