feat: Ask One More When ESC is Pressed
When you press down ESC while playing, a message pops up and asks
if you really want to come back to main menu.
To implement this, another scene overlappes the playing scene,
which means there would be two or more scenes for a while.
Also, this requires event handling logic to be sophisticated.