fix: Correct Failure When Server Is Full
Each listener of both servers tried to occupy null socket which didn't exist, after accepting a new connection with occupying the last socket.  This's now fixed.
When each server is full and then wins room for new connection, each listener starts accepting new connection.
Even if the queue server disconnected from the main server, the queue server can connect with the main server again while the main server works on, because in the main server the socket for the queue server is dealt with as a client like other game clients.
Thus, a term 'connection' above doesn't always indicate one for the game client.
A game client disconnected from the queue server while waiting for entry into the main server is now also removed out of the queue in the queue server.