fix: Run Successfully Even in Release Mode Build.
I figured out project Server was the culprit having made transfer too much slower.  At first, I assumed that project setting had an effect on slow transfer.  So it looked like that CRT was the key because transfer run fast with /MDd for project Server and /MD for others.  But that's a trap.  On the verge of getting stuck into map files, dumps or CRT debugging, it stroke me FPS of Server probably dropped and it was the true.  By profiling, I found out what shouldn't be sent every frame had been sent every frame but, to my surprise, slowness in debug mode build made it send at intervals as intended.